Update testcontainers-scala to v0.38.8
