+++ mysql/src/test/scala/zio/sql/TestContainer.scala
@@ -1,7 +1,8 @@
 package zio.sql
 
 import com.dimafeng.testcontainers.SingleContainer
-import com.dimafeng.testcontainers.PostgreSQLContainer
+import com.dimafeng.testcontainers.MySQLContainer
+import org.testcontainers.utility.DockerImageName
 import zio._
 import zio.blocking.{ effectBlocking, Blocking }
 
@@ -15,13 +16,11 @@ object TestContainer {
       }
     }(container => effectBlocking(container.stop()).orDie).toLayer
 
-  def postgres(
-    imageName: Option[String] = Some("postgres:alpine:13")
-  ): ZLayer[Blocking, Throwable, Has[PostgreSQLContainer]] =
+  def mysql(imageName: String): ZLayer[Blocking, Throwable, Has[MySQLContainer]] =
     ZManaged.make {
       effectBlocking {
-        val c = new PostgreSQLContainer(
-          dockerImageNameOverride = imageName
+        val c = new MySQLContainer(
+          mysqlImageVersion = Option(imageName).map(DockerImageName.parse)
         ).configure { a =>
           a.withInitScript("shop_schema.sql")
           ()
@@ -29,8 +28,6 @@ object TestContainer {
         c.start()
         c
       }
-    } { container =>
-      effectBlocking(container.stop()).orDie
-    }.toLayer
+    }(container => effectBlocking(container.stop()).orDie).toLayer
 
 }
